Add header resposne from server, resource error checking, 404 notification

+++ web_server/server.rb
@@ -9,22 +9,29 @@ loop {
   request = client.gets
   resource = request[/\w+.html/]
   type = request[/(GET|POST)/]
-
+  version = request[/HTTP\/\d\.\d/]
+  message = ""
+  status = ""
   response = ""
-  file = File.open(resource, "r")
-  file.each do |line|
-    response << line
-  end
+
+  puts "Request: " + request.chomp
 
   if type == "GET"
-    puts "Request: " + request + " Sending: #{resource}"
+    if File.exists?("#{resource}")
+      puts "Resource found, sending: #{resource}"
+      file = File.open(resource, "r")
+      file.each do |line|
+        response << line; end
+      status = "200"
+      message = "OK"
+    else
+      puts "Resource not found, sending code 404."
+      status = "404"
+      message = "Not Found"
+    end
+    client.puts(version + " " + status + " " + message)
     client.puts(response)
   end
 
-
-
-
-  client.puts(Time.now.ctime)
-  client.puts "Closing connection. "
   client.close
 }
